requirements.txt
certifi==2020.6.20
click==7.1.2
distro==1.5.0
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.1
pandas==1.1.0
psycopg2-binary==2.8.5
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.18
tabula==1.0.5
tabula-py==2.1.1
Werkzeug==1.0.1

pip~=20.2.3
cryptography~=3.0
py~=1.7.0
lxml~=4.2.5
keyring~=17.0.0
setuptools~=47.3.1
Cython~=0.29.2
requests~=2.20.1
bs4~=0.0.1
beautifulsoup4~=4.6.3
schedule~=0.6.0
python-dotenv~=0.14.0